+++ config/warnings.js
@@ -4,7 +4,9 @@
  */
 module.exports = {
   plugins: ['import'],
-  rules: { 'import/no-named-as-default': 1
-         , 'import/no-duplicates': 1
-         }
+  rules: {
+    'import/no-named-as-default': 1,
+    'import/no-named-as-default-member': 1,
+    'import/no-duplicates': 1,
+  },
 }